nouvelle modif

+++ routes/web.php
@@ -17,6 +17,7 @@ use App\Http\Controllers\ProduitController;
 use App\Http\Controllers\CommandeController;
 use App\Http\Controllers\CategorieController;
 use App\Http\Controllers\FournisseurController;
+use GuzzleHttp\Middleware;
 
 /*
 |--------------------------------------------------------------------------
@@ -30,26 +31,44 @@ use App\Http\Controllers\FournisseurController;
 */
 
 Route::get('/produit',[ProduitController::class,'index'])->name('produit.list');
+
 Route::get('/client',[ClientController::class,'index'])->name('client.liste');
-Route::get('/fournisseur',[FournisseurController::class,'index'])->name('fournisseur.liste');
-Route::get('/role',[UserController::class,'liste'])->name('role.liste');
-Route::get('/categorie',[CategorieController::class,'index'])->name('categorie.liste');
-Route::get('/type',[TypeController::class,'index'])->name('type.liste');
+
+Route::get('/fournisseur',[FournisseurController::class,'index'])->name('fournisseur.liste')->middleware('auth.admin');
+Route::get('/role',[UserController::class,'liste'])->name('role.liste')->middleware('auth.admin');
+Route::get('/categorie',[CategorieController::class,'index'])->name('categorie.liste')->middleware('auth.utilisateur');
+Route::get('/type',[TypeController::class,'index'])->name('type.liste')->middleware('auth.utilisateur');
 Route::get('/produits',[ProduitController::class,'create'])->name('product.create');
-Route::get('/users',[UserController::class,'index'])->name('user.create');
+Route::get('/users',[UserController::class,'index'])->name('user.create')->middleware('auth.admin');
 Route::get("client/create",[ClientController::class,'indes'])->name('client.create');
-Route::get("/fournisseur/create",[FournisseurController::class,'inde'])->name('fournisseur.create');
-Route::get('/contrats',[ContratController::class,'index'])->name('contrat');
+Route::get("/fournisseur/create",[FournisseurController::class,'inde'])->name('fournisseur.create')->middleware('auth.admin');
+Route::get('/contrats',[ContratController::class,'index'])->name('contrat')->middleware('auth.admin');
 Auth::routes();
 Route::get('/', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
-Route::get('/rayon',[RayonController::class,'index'])->name('rayon.index');
-Route::get('commade',[CommandeController::class,'index'])->name('commande');
-Route::get('/commandes/produits/{id}',[CommandeController::class,'commandeArticle'])->name('commande.article');
-Route::get('/commande/produit',[CommandeController::class,'view2'])->name('liste.commande');
-Route::get('/bon/commande/{id}',[CommandeController::class,'bonCommande'])->name('bon.commande');
-Route::get('/edit/commande/{id}',[CommandeController::class,'edition'])->name('edit.commande');
-Route::get('livraison',[CommandeController::class,'livraison'])->name('livraison');
-Route::get('/ventes',[VenteController::class,'index'])->name('vente');
+
+Route::get('/rayon',[RayonController::class,'index'])->name('rayon.index')->middleware('auth.utilisateur');
+Route::get('commade',[CommandeController::class,'index'])->name('commande')->middleware('auth.admin');
+
+
+Route::get('/commandes/produits/{id}',[CommandeController::class,'commandeArticle'])->name('commande.article')
+->middleware('auth.admin');
+
+Route::get('/commande/produit',[CommandeController::class,'view2'])->name('liste.commande')->middleware('auth.admin');
+
+
+Route::get('/bon/commande/{id}',[CommandeController::class,'bonCommande'])->name('bon.commande')
+->middleware('auth.admin');
+
+Route::get('/edit/commande/{id}',[CommandeController::class,'edition'])->name('edit.commande')
+->Middleware('auth.admin');
+
+Route::get('livraison',[CommandeController::class,'livraison'])->name('livraison')
+->middleware('auth.admin');
+
+
+Route::get('/ventes',[VenteController::class,'index'])->name('vente')->middleware('auth.utilisateur');
+
+
 Route::get('/vente/produit/{id}',[VenteController::class,'edit'])->name('vente.produit');
 Route::post('/vente/',[VenteController::class,'venteProduit'])->name('vent');
 Route::get('/liste/vente',[VenteController::class,'listeVente'])->name('liste.vente');
@@ -57,45 +76,52 @@ Route::get('/pdf/inventaire',[ProduitController::class,'inventairePDF'])->name('
 Route::get('autocomplete', [VenteController::class, 'autocomplete'])->name('autocomplete');
 /**route relatives aux enregistrements */
 Route::post('/client',[ClientController::class,'create'])->name('create.client');
-Route::Post('/fournisseur',[FournisseurController::class,'insertion'])->name('insert.fournisseur');
+Route::Post('/fournisseur',[FournisseurController::class,'insertion'])->name('insert.fournisseur')->middleware('auth.admin');
 Route::post('type',[TypeController::class,'create'])->name('insert.create');
 Route::post('categorie',[CategorieController::class,'create'])->name('categorie.create');
 
-Route::POST('/users',[UserController::class,'insertion'])->name('users.create');
+Route::POST('/users',[UserController::class,'insertion'])->name('users.create')->middleware('auth.admin');
 
-Route::get('fournisseur/pdf',[FournisseurController::class,'pdf'])->name('fournisseur.pdf');
+Route::get('fournisseur/pdf',[FournisseurController::class,'pdf'])->name('fournisseur.pdf')->middleware('auth.admin');
 route::post('produits',[ProduitController::class,'save'])->name('produit.create');
 Route::post('/commande',[CommandeController::class,'save'])->name('commande.produit');
-Route::post('commande/produit/liste',[CommandeController::class,'command'])->name('command');
-Route::put('/com/{commande}',[CommandeController::class,'edit'])->name('commande.edition');
+
+Route::post('commande/produit/liste',[CommandeController::class,'command'])->name('command')->middleware('auth.admin');
+Route::put('/com/{commande}',[CommandeController::class,'edit'])->name('commande.edition')->middleware('auth.admin');
 Route::post('/vente/index',[VenteController::class,'create'])->name('vente.index');
-Route::post('contrat/create',[ContratController::class,'store'])->name('contrat.create');
-Route::get('contrat/creation',[ContratController::class,'vue2'])->name('contrat.creation');
-route::get('contrat/edit/{id}',[ContratController::class,'update'])->name('edit');
+Route::post('contrat/create',[ContratController::class,'store'])->name('contrat.create')->middleware('auth.admin');
+
+
+Route::get('contrat/creation',[ContratController::class,'vue2'])->name('contrat.creation')->middleware('auth.admin');
+route::get('contrat/edit/{id}',[ContratController::class,'update'])->name('edit')->middleware('auth.admin');
 route::put('contrat/upgrade/{contrat}',[ContratController::class,'updat'])->name('update.contrat');
 route::get('/facture/vente/{id}',[VenteController::class,'facture'])->name('facture');
 Route::get('/statistics',[VenteController::class,'statistique'])->name('chart');
 Route::get('edit/vente/{id}',[VenteController::class,'updateVente'])->name('edit.vente');
-Route::get('/update/product/{id}',[ProduitController::class,'update'])->name('update.produit');
-Route::put('/update/produict/{produits}',[ProduitController::class,'updat'])->name('produit.updat');
+Route::get('/update/product/{id}',[ProduitController::class,'update'])->name('update.produit')->middleware('auth.admin');
+Route::put('/update/produict/{produits}',[ProduitController::class,'updat'])->name('produit.updat')->middleware('auth.admin');
 Route::put('/liste/vente/{vent}',[VenteController::class,'updateView'])->name('vente.update');
 
-route::get('/role/users/{id}',[UserController::class,'vue'])->name('role.vue');
-route::put('/role/user/{user}',[UserController::class,'upgrade'])->name('user.upgrade');
+route::get('/role/users/{id}',[UserController::class,'vue'])->name('role.vue')->middleware('auth.admin');
+route::put('/role/user/{user}',[UserController::class,'upgrade'])->name('user.upgrade')->middleware('auth.admin');
 
-Route::delete('/liste/vente/{id}',[VenteController::class,'deleteVente'])->name('delete.vente');
-Route::get('bon/livraison/{id}',[CommandeController::class,'bonlivraison'])->name('bon.livraison');
+Route::delete('/liste/vente/{id}',[VenteController::class,'deleteVente'])->name('delete.vente')->middleware('auth.admin');
+Route::get('bon/livraison/{id}',[CommandeController::class,'bonlivraison'])->name('bon.livraison')->middleware('auth.admin');
 Route::get('search-from-db', [VenteController::class, 'searchDB']);
-Route::get('/fournisseur/edit/{id}',[FournisseurController::class,'edit'])->name('fournisseur.edit');
-Route::delete('/contrats/delete/{id}',[ContratController::class,'archiver'])->name('contrat.delete');
-Route::put('/fournisseur/edition/{fournisseur}',[FournisseurController::class,'editer'])->name('fournisseur.edition');
+Route::get('/fournisseur/edit/{id}',[FournisseurController::class,'edit'])->name('fournisseur.edit')->middleware('auth.admin');
+Route::delete('/contrats/delete/{id}',[ContratController::class,'archiver'])->name('contrat.delete')->middleware('auth.admin');
+Route::put('/fournisseur/edition/{fournisseur}',[FournisseurController::class,'editer'])->name('fournisseur.edition')->middleware('auth.admin');
 Route::get('/changePassword',[HomeController::class,'showChangePasswordForm']);
 Route::post('/changePassword',[HomeController::class,'changePassword'])->name('changePassword');
 // route::get('/liste/vente/',[VenteController::class,'search'])->name('search.vente');
 
 
-Route::get('caisse',[CaisseController::class,'index'])->name('caisse.index');
-Route::get('/caisse/etat',[CaisseController::class,'etat'])->name('etat');
+Route::get('caisse',[CaisseController::class,'index'])->name('caisse.index')
+->Middleware('auth.utilisateur');
+
+Route::get('/caisse/etat',[CaisseController::class,'etat'])->name('etat')
+->middleware('auth.utilisateur');
+
 Route::post('/daterange/fetch_data', [CaisseController::class, 'fetch_data'])->name('daterange.fetch_data');
 
 Route::get('/commande/produit/{id}', [CommandeController::class, 'addToCart'])->name('add.to.cart');
@@ -110,8 +136,8 @@ route::get('/vente/group',[VenteController::class,'Ventegroup'])->name('vente.gr
 route::get('facture/group',[VenteController::class,'factureGroupe'])->name('group.facture');
 
 route::get('/information',[HomeController::class,'info'])->name('infos');
-route::post('rayon/create',[RayonController::class,'newCreate'])->name('rayon.create');
 
+route::post('rayon/create',[RayonController::class,'newCreate'])->name('rayon.create');
 Route::patch('update-cart', [VenteController::class, 'updated'])->name('update.cart');
 Route::delete('/enlever',[VenteController::class,'remov'])->name('retirer');
 
@@ -119,8 +145,6 @@ Route::delete('/enlever',[VenteController::class,'remov'])->name('retirer');
 /**nouvele carte */
 Route::post('clear', [VenteController::class, 'clearAllCart'])->name('cart.clear');
 Route::post('cart', [VenteController::class, 'addToCart'])->name('cart.store');
-
-
 Route::post('cart/add',[CommandeController::class,'addCart'])->name('cart.add');
 Route::post('/cart/clean',[CommandeController::class,'clearAllCart'])->name('clear.commande');
 Route::post('Group/livrable',[CartController::class,'addLivraison'])->name('cart.livraison');
